Refactored Expense and NewItemView (-> yang ini ga semua bisa di jadiin custom component gara2 focusState Bindingnya gw gak paham :\  )
