+++ MC1GD/Views/NewItemView.swift
@@ -9,6 +9,7 @@ import SwiftUI
 import Combine
 let dateNotif = PassthroughSubject<Date, Never>()
 struct NewItemView: View {
+    private let categories = ["Makanan dan Minuman", "Transportasi", "Barang"]
     @Binding var showSheet : Bool
     @StateObject var formVm = addItemFormViewModel()
     @State var newItemImage = UIImage()
@@ -16,7 +17,6 @@ struct NewItemView: View {
     @State var newItemCategory = ""
     @State var newItemTag : String = ""
     @State var newItemDesc : String = ""
-    private let categories = ["Makanan dan Minuman", "Transportasi", "Barang"]
     @EnvironmentObject var viewModel : coreDataViewModel
     @FocusState var isFocusedName : Bool
     @FocusState var isFocusedPrice : Bool
@@ -44,56 +44,7 @@ struct NewItemView: View {
                     // MARK: harga barang
                     Group {
                         HStack{
-                            Text("Rp")
-                                .font(.caption)
-                                .fontWeight(.semibold)
-                                .padding(.horizontal,10)
-                                .padding(.vertical,6)
-                                .background(
-                                    RoundedRectangle(cornerRadius: 10)
-                                        .stroke(Color.gray)
-                                        .background(Color("tertiary-gray").cornerRadius(10)))
-                                .foregroundColor(Color("primary-gray"))
-                                .frame(width: 40)
-                            Divider().frame(height: 30)
-                            TextField("Harga Barang", text: $newItemPrice)
-                                .disableAutocorrection(true)
-                                .lineLimit(0)
-                                .keyboardType(.numberPad)
-                                .font(.title2)
-                                
-                                .onChange(of: newItemPrice) { newValue in
-                                    withAnimation {
-                                        let numberString = newValue.replacingOccurrences(of: "[^0-9]", with: "", options: .regularExpression)
-                                        if let number = Double(numberString) {
-                                            num = number
-                                            newItemPrice = currencyFormatterInForm.string(from: NSNumber(value: num)) ?? ""
-                                        }
-                                        if (numberString != "0" && numberString != ""){
-                                            priceValid = true
-                                        }
-                                        else{
-                                            priceValid = false
-                                        }
-                                    }
-                                    
-                                }
-                                .overlay (
-                                    ZStack{
-                                        Button {
-                                            withAnimation {
-                                                newItemPrice = ""
-                                            }
-                                        } label: {
-                                            Image(systemName: "delete.left").foregroundColor(.red)
-                                        }
-                                        .buttonStyle(.borderless)
-                                        .clipped()
-                                        .opacity(priceValid ? 1.0 : 0.0)
-                                        
-                                    }
-                                    , alignment: .trailing
-                                )
+                            ItemPriceTextField(newItemPrice: $newItemPrice, num: $num, priceValid: $priceValid)
                                 .focused($focusedField, equals: Field.itemPrice)
                                 .onTapGesture {
                                     focusedField = Field.itemPrice
@@ -113,34 +64,13 @@ struct NewItemView: View {
                                 .padding(.horizontal,15)
                         }
                     }
-                    
-                    
+
                     Divider()
                     
                     // MARK: nama barang
                     Group {
                         HStack{
-                            Text("Aa")
-                                .font(.headline)
-                                .foregroundColor(Color.primary_gray)
-                                .padding(.horizontal,15)
-                            TextField("Nama Barang", text: $formVm.itemName).autocorrectionDisabled(true)
-                                .overlay (
-                                    ZStack{
-                                        Button {
-                                            withAnimation {
-                                                formVm.deleteText()
-                                            }
-                                        } label: {
-                                            Image(systemName: "delete.left").foregroundColor(.red)
-                                        }
-                                        .buttonStyle(.borderless)
-                                        .clipped()
-                                        .opacity(formVm.buttonDeleteOn ? 1.0 : 0.0)
-                                        
-                                    }
-                                    , alignment: .trailing
-                                )
+                            itemNameTextField(formVm: formVm)
                                 .focused($focusedField, equals: Field.itemName)
                                 .onTapGesture {
                                     focusedField = Field.itemName
@@ -163,116 +93,13 @@ struct NewItemView: View {
                     
                     // MARK: pilih itemCategory
                     Group {
-                        HStack{
-                            Image(systemName: "circle.fill")
-                                .imageScale(.large)
-                                .foregroundColor(Color.primary_gray)
-                                .padding(.horizontal,15)
-                            HStack {
-                                Picker("", selection: $newItemCategory) {
-                                    ForEach(categories, id: \.self){ category in
-                                        Text(category)
-                                    }
-                                }
-                                .pickerStyle(.menu)
-                                .tint(Color.primary)
-                                .clipped()
-                            }
-                        }.padding(.vertical, 5)
-                        Divider()
+                        itemCategoryPicker(newItemCategory: $newItemCategory)
                     }
                     
                     
                     // MARK:  pilih itemTag needs/wants
                     Group {
-                        VStack (alignment: .leading){
-                            HStack{
-                                Image(systemName: "tag")
-                                    .imageScale(.large)
-                                    .foregroundColor(Color.primary_gray)
-                                    .padding(.horizontal,15)
-                                Button{
-                                    if isNeeds != true {
-                                        if isWants != true {
-                                            self.isWants.toggle()
-                                            self.newItemTag = "Keinginan"
-                                        }else{
-                                            self.isWants.toggle()
-                                            self.newItemTag = ""
-                                        }
-                                    }else{
-                                        self.isNeeds.toggle()
-                                        self.isWants.toggle()
-                                        self.newItemTag = "Keinginan"
-                                    }
-                                } label: {
-                                    Text("Keinginan")
-                                        .font(.caption)
-                                        .bold()
-                                        .padding(2)
-                                        .textCase(.uppercase)
-                                    
-                                }
-                                .buttonStyle(.bordered)
-                                .foregroundColor(.white)
-                                .background(
-                                    Color.tag_pink
-                                        .opacity(self.isWants ? 1: 0.4)
-                                )
-                                .cornerRadius(25)
-                                .hoverEffect(.lift)
-                                Button{
-                                    if isWants != true{
-                                        if isNeeds != true {
-                                            self.isNeeds.toggle()
-                                            self.newItemTag = "Kebutuhan"
-                                        }else{
-                                            self.isNeeds.toggle()
-                                            self.newItemTag = ""
-                                        }
-                                    }else{
-                                        self.isWants.toggle()
-                                        self.isNeeds.toggle()
-                                        self.newItemTag = "Kebutuhan"
-                                    }
-                                } label: {
-                                    Text("Kebutuhan")
-                                        .font(.caption)
-                                        .bold()
-                                        .padding(2)
-                                        .textCase(.uppercase)
-                                    
-                                }
-                                .buttonStyle(.bordered)
-                                .foregroundColor(.white)
-                                .background(
-                                    Color.tag_purple
-                                        .opacity(self.isNeeds ? 1: 0.4)
-                                )
-                                .cornerRadius(25)
-                                .hoverEffect(.lift)
-                                
-                                Button{
-                                    showQuestions.toggle()
-                                } label: {
-                                    Image(systemName: "questionmark.circle.fill")
-                                        .foregroundColor(Color("secondary-gray"))
-                                }
-                                .buttonStyle(.borderless)
-                                .sheet(isPresented: $showQuestions) {
-                                    QuestionsView(showQuestions: $showQuestions)
-                                }
-                                
-                                
-                            }
-                            Text("Pilih Salah Satu")
-                                .font(.caption2)
-                                .padding(.leading,75)
-                                .italic()
-                                .foregroundColor(Color.secondary_gray)
-                        }
-                        .padding(.top,5)
-                        Divider()
+                        itemTagField(isNeeds: $isNeeds, isWants: $isWants, showQuestions: $showQuestions, newItemTag: $newItemTag)
                     }
                     
                     
@@ -310,33 +137,7 @@ struct NewItemView: View {
                     
                     // MARK: select tanggal
                     Group{
-                        HStack{
-                            Image(systemName: "calendar")
-                                .imageScale(.large)
-                                .foregroundColor(Color("secondary-gray"))
-                                .padding(.horizontal,15)
-                            
-                            Button{
-                                showDatePicker.toggle()
-                            }label:{
-                                Text(Date().formatDateFull(for: todayDateComponent))
-                                    .font(.body)
-                                    .padding(.vertical,8)
-                                    .padding(.horizontal,10)
-                                    .foregroundColor(.black)
-                                    .background(
-                                        RoundedRectangle(cornerRadius: 10)
-                                            .stroke(Color.tertiary_gray)
-                                    )
-                            }
-                            
-                        }.padding(.top,10)
-                        if showDatePicker == true{
-                            HStack{
-                                DatePicker("" ,selection: $todayDateComponent, in: ...Date(), displayedComponents: .date).datePickerStyle(.wheel)
-                                    .accentColor(Color.primary_purple)
-                            }
-                        }
+                        datePickerField(showDatePicker: $showDatePicker, todayDateComponent: $todayDateComponent)
                     }
                 }
                 .padding(.horizontal,20)
@@ -344,13 +145,11 @@ struct NewItemView: View {
                 .onTapGesture{
                     focusedField = nil
                 }
-                
                 Spacer()
-                
-                
             }
             .onAppear{
-                newItemCategory = categories.first!}
+                newItemCategory = categories.first!
+            }
             .toolbar{
                 ToolbarItem(placement: .navigationBarLeading) {
                     Button("Kembali"){
